A bit of a reorganisation
 - Move all executables into the "bin" folder.
 - Move test-related code into "tests/driver".
