+++ bin/GenExample.hs
