+++ bin/Version.hs
