+++ tests/driver/Test/Frontend/Amc.hs
@@ -1,7 +1,7 @@
 module Test.Frontend.Amc (tests) where
 
+import Test.Tasty.Cram
 import Test.Tasty
-import Test.Cram
 
 tests :: IO TestTree
 tests = testGroup "amc" <$> cramDir "tests/amc"
