+++ tests/driver/Test/Syntax/Verify.hs
